+++ pkg/cmd/create/authrealm/exec.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 
 	clusteradmhelpers "open-cluster-management.io/clusteradm/pkg/helpers"
-	clusteradmapply "open-cluster-management.io/clusteradm/pkg/helpers/apply"
 
+	"github.com/stolostron/applier/pkg/apply"
 	"github.com/stolostron/cm-cli/pkg/cmd/create/authrealm/scenario"
 	"github.com/stolostron/cm-cli/pkg/helpers"
 	apiextensionsclient "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
@@ -130,7 +130,7 @@ func (o *Options) runWithClient(kubeClient kubernetes.Interface,
 	output := make([]string, 0)
 
 	reader := scenario.GetScenarioResourcesReader()
-	applierBuilder := clusteradmapply.NewApplierBuilder()
+	applierBuilder := apply.NewApplierBuilder()
 	applier := applierBuilder.WithClient(kubeClient, apiextensionsClient, dynamicClient).Build()
 
 	iauthRealm := o.values["authRealm"]
@@ -214,5 +214,5 @@ func (o *Options) runWithClient(kubeClient kubernetes.Interface,
 	}
 	output = append(output, out...)
 
-	return clusteradmapply.WriteOutput(o.outputFile, output)
+	return apply.WriteOutput(o.outputFile, output)
 }